requirements.txt
Appium-Python-Client~=2.11.1
pytest~=8.3.5
urllib3~=2.4.0
selenium~=4.32.0
allure-pytest~=2.1.0